pages-editor-2025-pt4: reorganise task types folder
