+++ app/pages/lab-pages-editor/components/TasksPage/components/EditStepDialog/task-types/DrawingTask.jsx
